REVIEW NEEDED: Implement machinery to interrupt mid-scan
Introduce a new global variable, `stop_flag`,
periodically checked by the scan handlers during the scan.
If it's true, the scan is interrupted and only matches obtained
till that point are returned.

For front-end use, introduce the public API function
`sm_set_stop_flag(bool)` to set the flag,
so that a scan stop can be called at anytime by the front-end.