libscanmem: Completely remove void pointer arithmetic
void* arithmetic is a GNU extension and uses a sizeof of 1.

Replace all the void pointers that need arithmetic done on
them with char* pointers by default, to obtain the same result.

Also change the return type of the `local_address...()` functions in
`targetmem` to OVaMI*, as they would naturally be, and introduce
explicit casts when they need to be treated as swaths or
used for bytes arithmetic.

Speed of any operation unchanged.